% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PA.EMOA.dataprocessing.R
\name{addAllGroup}
\alias{addAllGroup}
\title{Generate a new group that gathers all other groups.}
\usage{
addAllGroup(df, col, group = "all")
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data frame.}

\item{col}{[\code{character(1)}]\cr
Column (factor or character) used to generate \dQuote{all}-group.}

\item{group}{[\code{character(1)}]\cr
New group name.
Default is \dQuote{all}.}
}
\value{
[\code{data.frame}] Modified input \code{df}
}
\description{
Given a data frame and a column name all rows of the data frame
are duplicated and stored under a new group with respect to the column.
}
