% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PA.EMOA.applyStatisticalTests.R
\name{applyStatisticalTests}
\alias{applyStatisticalTests}
\title{Statistical tests of EMOA indicators.}
\usage{
applyStatisticalTests(inds, ind.names, alpha = 0.05, ...)
}
\arguments{
\item{inds}{[\code{data.frame}]\cr
Input data frame (return value of \code{\link{computeIndicators}}).}

\item{ind.names}{[\code{character}]\cr
Optional character vector of indicators (column names of \code{inds}) which
shall be considered in tests.
Defaults to all indicators.}

\item{alpha}{[\code{numeric(1)}]\cr
Siginificance level for statistical tests.
Default is 0.05.}

\item{...}{[any]\cr
Not used at the moment.}
}
\value{
[\code{list}] Named list of lists of matrizes. The names of the top level
  are the names of the problems (\code{unique(inds$prob)}), the names
  on the second level are the names of the considered indicators (\code{ind.names}).
  Each component of the matrizes on the bottom level contains the adjusted
  \eqn{p}-values of the corresponding location test of the indicator for two
  algorithms (row- and column names of the matrix).
}
\description{
Given a data.frame as returned by \code{\link{computeIndicators}},
the function performs pairwise tests (non-parameteric Wilcoxon sum test at the
moment) for location differences in the indicator distributions for all
algorithms. Multiple testing issues are handled internally as is the \dQuote{direction}
of the indicators (e.g., the unary hypervolume indicator shall be minimized
while the number of non-dominated solutions is to be maximized).
}
\references{
[1] Knowles, J., Thiele, L., & Zitzler, E. (2006). A Tutorial on the Performance Assessment
of Stochastic Multiobjective Optimizers. Retrieved from https://sop.tik.ee.ethz.ch/KTZ2005a.pdf
}
\seealso{
Other EMOA performance assessment tools: \code{\link{approximateNadirPoint}},
  \code{\link{computeDominanceRanking}},
  \code{\link{emoaIndEps}},
  \code{\link{makeEMOAIndicator}},
  \code{\link{niceCellFormater}}, \code{\link{normalize}},
  \code{\link{plotDistribution}}, \code{\link{plotFront}},
  \code{\link{plotScatter2d}}, \code{\link{plotScatter3d}},
  \code{\link{toLatexTables}}
}
