% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCrowdingDistance.R
\name{computeCrowdingDistance}
\alias{computeCrowdingDistance}
\title{Compute the crowding distance of a set of points.}
\usage{
computeCrowdingDistance(x)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Numeric matrix with each column representing a point.}
}
\value{
[\code{numeric}] Vector of crowding distance values.
}
\description{
The crowding distance is a measure of spread of solutions in the
approximation of the Pareto front. It is used, e.g., in the NSGA-II algorithm
as a second selection criterion.
}
\references{
K. Deb, A. Pratap, S. Agarwal, T. Meyarivan, A fast and elitist
multiobjective genetic algorithm: NSGA-II, IEEE Transactions on Evolutionary
Computation In Evolutionary Computation, IEEE Transactions on, Vol. 6, No. 2.
(07 April 2002), pp. 182-197, doi:10.1109/4235.996017
}
