% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAverageHausdorffDistance.R
\name{computeDistanceFromPointToSetOfPoints}
\alias{computeDistanceFromPointToSetOfPoints}
\title{Computes distance between a single point and set of points.}
\usage{
computeDistanceFromPointToSetOfPoints(
  a,
  B,
  dist.fun = computeEuclideanDistance
)
}
\arguments{
\item{a}{[\code{numeric(1)}]\cr
Point given as a numeric vector.}

\item{B}{[\code{matrix}]\cr
Point set (each column corresponds to a point).}

\item{dist.fun}{[\code{matrix}]\cr
Distance function to compute distance between points x and y. Expects a single
numeric vector d with the coordinate-wise differences di = (xi - yi).
Default is \code{computeEuclideanDist}.}
}
\value{
[\code{numeric(1)}]
}
\description{
Helper to compute distance between a single point and a point set.
}
