% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominates.R
\name{dominates}
\alias{\%dominates\%}
\alias{\%isDominatedBy\%}
\alias{dominates}
\alias{isDominated}
\title{Dominance relation check.}
\usage{
dominates(x, y)

isDominated(x, y)

x \%dominates\% y

x \%isDominatedBy\% y
}
\arguments{
\item{x}{[\code{numeric}]\cr
First vector.}

\item{y}{[\code{numeric}]\cr
Second vector.}
}
\value{
[\code{logical(1)}]
}
\description{
Check if a vector dominates another (\code{dominates}) or is
dominated by another (\code{isDominated}). There are corresponding infix
operators \code{dominates} and \code{isDominatedBy}. Moreover, function
\code{isMaximallyDominated} checks, whether a point is located in the last
layer of nondominated sorting algorithm.
}
\keyword{optimize}

