% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFitness.R
\name{evaluateFitness}
\alias{evaluateFitness}
\title{Computes the fitness value(s) for each individual of a given set.}
\usage{
evaluateFitness(control, inds, ...)
}
\arguments{
\item{control}{[\code{ecr_control}]\cr
Control object.}

\item{inds}{[\code{list}]\cr
List of individuals.}

\item{...}{[any]\cr
Optional parameters passed down to fitness function.}
}
\value{
[\code{matrix}].
}
\description{
This function expects a list of individuals, computes the fitness and always
returns a matrix of fitness values; even in single-objective optimization a
(1 x n) matrix is returned for consistency, where n is the number of individuals.
This function makes use of \code{\link[parallelMap]{parallelMap}} to
parallelize the fitness evaluation.
}
