% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.utils.R
\name{filterDuplicated}
\alias{filterDuplicated}
\title{Filter approximation sets by duplicate objective vectors.}
\usage{
filterDuplicated(res)
}
\description{
Filter approximation sets by duplicate objective vectors.
}
\note{
Note that this may be misleading if there can be solutions with identical
objective function values but different values in decision space.
}
