% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOffspring.R
\name{generateOffspring}
\alias{generateOffspring}
\title{Creates list of offspring individuals}
\usage{
generateOffspring(control, population, fitness, lambda, p.recomb = 0.7,
  p.mut = 0.1)
}
\arguments{
\item{control}{[\code{ecr2_control}]\cr
Control object.}

\item{population}{[\code{list}]\cr
Current population, i.e., list of individuals.}

\item{fitness}{[\code{matrix}]\cr
Matrix of fitness values with one column per individual of \code{individuals}.}

\item{lambda}{[\code{integer(1)}]\cr
Number of individuals generated in each generation.}

\item{p.recomb}{[\code{numeric(1)}]\cr
Probability of two parents to perform crossover.
Default is 0.7.}

\item{p.mut}{[\code{numeric(1)}]\cr
Probability to apply mutation to a child.
Default is 0.1.}
}
\value{
[\code{list}] Offspring.
}
\description{
Given and optimization state and a mating pool of individuals this function
generates offspring individuals based on the parameters specified in the
control object.
}

