% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRecombinator.R
\name{generatesMultipleChildren}
\alias{generatesMultipleChildren}
\title{Does the recombinator generate multiple children?}
\usage{
generatesMultipleChildren(recombinator)
}
\arguments{
\item{recombinator}{[\code{function}]\cr
Actual mutation operator.}
}
\value{
[\code{logical}]
  Boolean
}
\description{
Returns as to whether the recombinator generates multiple Children.
}
