% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getFront}
\alias{getFront}
\title{Extract fitness values from Pareto archive.}
\usage{
getFront(x)
}
\arguments{
\item{x}{[\code{ecr_pareto_archive}]\cr
Pareto archive.}
}
\value{
[\code{matrix}]
}
\description{
Get all non-dominated points in objective space, i.e., an (m x n)
matrix of fitness with m being the number of objectives and n being the number
of non-dominated points in the Pareto archive.
}
