% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getIndividuals}
\alias{getIndividuals}
\title{Extract individuals from Pareto archive.}
\usage{
getIndividuals(x)
}
\arguments{
\item{x}{[\code{ecr_pareto_archive}]\cr
Pareto archive.}
}
\value{
[\code{list}]
}
\description{
Get the non-dominated individuals logged in the Pareto archive.
}
\seealso{
Other ParetoArchive: \code{\link{getSize}},
  \code{\link{initParetoArchive}},
  \code{\link{updateParetoArchive}}
}
