% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRecombinator.R
\name{getNumberOfChildren}
\alias{getNumberOfChildren}
\title{Number of children}
\usage{
getNumberOfChildren(recombinator)
}
\arguments{
\item{recombinator}{[\code{function}]\cr
Actual mutation operator.}
}
\value{
[\code{numeric}]
  Number of children generated
}
\description{
Returns the number children generated by the recombinator
}
