% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRecombinator.R
\name{getNumberOfParentsNeededForMating}
\alias{getNumberOfParentsNeededForMating}
\title{Number of parents  needed for mating}
\usage{
getNumberOfParentsNeededForMating(recombinator)
}
\arguments{
\item{recombinator}{[\code{function}]\cr
Actual mutation operator.}
}
\value{
[\code{numeric}]
  Number of Parents need for mating
}
\description{
Returns the number of parents needed for mating.
}
