% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{getOperatorName}
\alias{getOperatorName}
\title{Get name of a operator.}
\usage{
getOperatorName(operator)
}
\arguments{
\item{operator}{[\code{ecr_operator}]\cr
Operator object.}
}
\value{
[\code{character(1)}]
  Name of the operator.
}
\description{
Returns the name of the passed operator.
}

