% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{getOperatorParameters}
\alias{getOperatorParameters}
\title{Get the operator's parameters.}
\usage{
getOperatorParameters(operator)
}
\arguments{
\item{operator}{[\code{ecr_operator}]\cr
Operator object.}
}
\value{
[\code{list}]
  Named list of parameters.
}
\description{
Return a named list of parameters the operator was initialized with.
}

