% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.default.R
\name{getPopulationFitness}
\alias{getPopulationFitness}
\title{Access to logged population fitness.}
\usage{
getPopulationFitness(log, trim = TRUE)
}
\arguments{
\item{log}{[\code{ecr_logger}]\cr
The log generated by \code{initLogger}.}

\item{trim}{[\code{logical(1)}]\cr
Should unused lines in the logged be cut off?
Usually one wants this behaviour.
Thus the default is \code{TRUE}.}
}
\value{
[\code{list}] List of populations.
}
\description{
Returns the fitness values of all individuals as a data.frame
with columns f1, ..., fo, where o is the number of objectives and column
\dQuote{gen} for generation.
}
\seealso{
Other logging: 
\code{\link{getPopulations}()},
\code{\link{getStatistics}()},
\code{\link{initLogger}()},
\code{\link{updateLogger}()}
}
\concept{logging}
