% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.default.R
\name{getPopulations}
\alias{getPopulations}
\title{Access to logged populations.}
\usage{
getPopulations(log, trim = TRUE)
}
\arguments{
\item{log}{[\code{ecr_logger}]\cr
The log generated by \code{initLogger}.}

\item{trim}{[\code{logical(1)}]\cr
Should unused lines in the logged be cut off?
Usually one wants this behaviour.
Thus the default is \code{TRUE}.}
}
\value{
[\code{list}] List of populations.
}
\description{
Simple getter for the logged populations.
}
\details{
This functions throws an error if the logger was initialized with
\code{log.pop = FALSE} (see \code{initLogger}).
}
\seealso{
Other logging: 
\code{\link{getPopulationFitness}()},
\code{\link{getStatistics}()},
\code{\link{initLogger}()},
\code{\link{updateLogger}()}
}
\concept{logging}
