% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getSize}
\alias{getSize}
\title{Get size of Pareto-archive.}
\usage{
getSize(x)
}
\arguments{
\item{x}{[\code{ecr_pareto_archive}]\cr
Pareto archive.}
}
\value{
[\code{integer(1)}]
}
\description{
Returns the number of stored individuals in Pareto archive.
}
\seealso{
Other ParetoArchive: \code{\link{getIndividuals}},
  \code{\link{initParetoArchive}},
  \code{\link{updateParetoArchive}}
}
