% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.default.R
\name{getStatistics}
\alias{getStatistics}
\title{Access the logged statistics.}
\usage{
getStatistics(log, trim = TRUE)
}
\arguments{
\item{log}{[\code{ecr_logger}]\cr
The log generated by \code{initLogger}.}

\item{trim}{[\code{logical(1)}]\cr
Should unused lines in the logged be cut off?
Usually one wants this behaviour.
Thus the default is \code{TRUE}.}
}
\value{
[\code{data.frame}] Logged statistics.
}
\description{
Simple getter for the logged fitness statistics.
}
\seealso{
Other logging: 
\code{\link{getPopulationFitness}()},
\code{\link{getPopulations}()},
\code{\link{initLogger}()},
\code{\link{updateLogger}()}
}
\concept{logging}
