% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{getSupportedRepresentations}
\alias{getSupportedRepresentations}
\title{Get supported representations.}
\usage{
getSupportedRepresentations(operator)
}
\arguments{
\item{operator}{[\code{ecr_operator}]\cr
Operator object.}
}
\value{
[\code{character}]
  Vector of representation types.
}
\description{
Returns the character vector of representation which the operator supports.
}
