% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.default.R
\name{initLogger}
\alias{initLogger}
\title{Initialize a log object.}
\usage{
initLogger(control, log.stats = list(fitness = list("min", "mean", "max")),
  log.pop = FALSE, init.size = 1000L)
}
\arguments{
\item{control}{[\code{ecr2_control}]\cr
Control object.}

\item{log.stats}{[\code{list}]\cr
List of lists for statistic computation on attributes of the individuals
of the population. Each entry should be named by the attribute it should be
based on, e.g., fitness, and should contain a list of R functions as a
character string or a a list with elements \code{fun} for the function, and \code{pars} for additional
parameters which shall be passed to the corresponding function.
Each function is required to return a scalar numeric value.
By default the minimum, mean and maximum of the fitness values is computed.}

\item{log.pop}{[\code{logical(1)}]\cr
Shall the entire population be saved?
Default is \code{FALSE}.}

\item{init.size}{[\code{integer(1)}]\cr
Initial number of rows of the stats slot of the logger, where the fitness
statistics are stored. The size of the statistics log doubled each time an
overflow occurs.
Default is 1000.}
}
\value{
[\code{ecr2_logger}]
  An S3 object of class \code{ecr2_logger} with the following components:
  \describe{
    \item{log.stats}{The \code{log.stats} list.}
    \item{log.pop}{The \code{log.pop} parameter.}
    \item{init.size}{Initial size of the log.}
    \item{env}{The actual log. This is an R environment which ensures, that
    in-place modification is possible.}
  }
}
\description{
Logging is a central aspect of each EA. Besides the final solution(s)
especially in research often we need to keep track of different aspects of the
evolutionary process, e.g., fitness statistics. The logger of \pkg{ecr} keeps
track of different user-defined statistics and the population.
It is also used for monitoring by the default console-monitor (see {makeECRConsoleMonitor})
and passed down to stopping conditions (see \code{makeECRTerminator}). Most
important this logger is used internally by the \code{ecr} black-box function.
}
\note{
If rapid prototyping of EAs is the goal and \code{ecr} is not sufficient
one is of cource not required to use this logger.
}
\examples{

}
\seealso{
Other logging: \code{\link{getLoggedPopulations}},
  \code{\link{getLoggedStats}}, \code{\link{updateLogger}}
}

