% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{is.supported}
\alias{is.supported}
\title{Check if ecr operator supports given representation.}
\usage{
is.supported(operator, representation)
}
\arguments{
\item{operator}{[\code{ecr_operator}]\cr
Object of type \code{ecr_operator}.}

\item{representation}{[\code{character(1)}]\cr
Representation as a string.}
}
\value{
[\code{logical(1)}]
  \code{TRUE}, if operator supports the representation type.
}
\description{
Check if the given operator supportds a certain representation, e.g., \dQuote{float}.
}

