% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{isEcrOperator}
\alias{isEcrOperator}
\title{Check if given function is an ecr operator.}
\usage{
isEcrOperator(obj)
}
\arguments{
\item{obj}{[any]\cr
Arbitrary R object to check.}
}
\value{
[\code{logical(1)}]
}
\description{
Checks if the passed object is of type \code{ecr_operator}.
}
