% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.makeEMOAIndicator.R
\name{makeEMOAIndicator}
\alias{makeEMOAIndicator}
\title{Constructor for EMOA indicators.}
\usage{
makeEMOAIndicator(fun, minimize, name, latex.name)
}
\arguments{
\item{fun}{[\code{function(points, ...)}]\cr
Function which expects a numeric matrix \dQuote{points} as first argument.
Optional named arguments (often \dQuote{ref.point} for a reference
point or \dQuote{ref.points} for a reference set, e.g., the true Pareto-front)
are allowed. See implementations of existing indicators for examples.}

\item{name}{[\code{character(1)}]\cr
Short name of the indicator. Used, e.g., as column name for the indicator
in the data.frame returned by \code{\link{computeIndicators}}.}

\item{latex.name}{[\code{character(1)}]\cr
LaTeX representation of the indicator. Used in LaTeX-table output statistical
tests (see \code{\link{toLatex}}).}

\item{ninimize}{[\code{logical(1)}]\cr
Lower values indicate better performance?}
}
\value{
[\code{function(points, ...)}] Argument \code{fun} with all other
  arguments appended.
}
\description{
Simple wrapper for function which compute
performance indicators for multi-objective stochastic algorithm.
Basically this function appends some meta information to the
passed function \code{fun}.
}
