% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOperator.R
\name{makeOperator}
\alias{makeOperator}
\title{Construct evolutionary operator.}
\usage{
makeOperator(operator, supported = getAvailableRepresentations())
}
\arguments{
\item{operator}{[\code{function}]\cr
Actual operator.}

\item{supported}{[\code{character}]\cr
Vector of names of supported parameter representations. Possible choices:
\dQuote{permutation}, \dQuote{float}, \dQuote{binary} or \dQuote{custom}.}
}
\value{
[\code{ecr_operator}] Operator object.
}
\description{
Helper function which constructs an evolutionary operator.
}
\note{
In general you will not need this function, but rather one of its
deriviatives like \code{\link{makeMutator}} or \code{\link{makeSelector}}.
}
