% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTerminator.R
\name{makeTerminator}
\alias{makeTerminator}
\title{Generate stopping condition.}
\usage{
makeTerminator(condition.fun, name, message)
}
\arguments{
\item{condition.fun}{[\code{function}]\cr
Function which takes a logger object \code{log} (see \code{\link{initLogger}})
and returns a single logical.}

\item{name}{[\code{character(1)}]\cr
Identifier for the stopping condition.}

\item{message}{[\code{character(1)}]\cr
Message which should be stored in the termination object, if the stopping
condition is met.}
}
\value{
[\code{ecr_terminator}]
}
\description{
Wrap a function within a stopping condition object.
}
