% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOffspring.R
\name{mutate}
\alias{mutate}
\title{Mutate helper.}
\usage{
mutate(x, inds, p.mut, par.list = list(), ...)
}
\arguments{
\item{x}{[\code{ecr2_control} | \code{ecr2_mutator}]\cr
Either a mutator or a control object. If a control object is passed
the functions is internally called on the mutator in \code{control$mutate}.
In this case the \code{control$params} list is passed down via \code{par.list}.}

\item{inds}{[\code{list}]\cr
List of individuals.}

\item{p.mut}{[\code{numeric(1)}]\cr
Probability to apply mutation to a child.
Default is 0.1.}

\item{par.list}{[\code{list}]\cr
Named list of parameters passed down to mutator.
Default is the empty list.}

\item{...}{[any]\cr
Further parameters passed down to mutators.
There parameters will overwrite parameters in \code{par.list}.}
}
\value{
[\code{list}] Mutated \code{inds} list.
}
\description{
Receives a list of individuals and a control object and applies
the mutation operator stored in control to each individual with a certain
probability.
}
\examples{
# generate list of bistrings and apply mutation
inds = replicate(5, sample(c(0, 1), 10, replace = TRUE), simplify = FALSE)
mutator = setupBitflipMutator(p = 0.1)
mut.inds = mutate(mutator, inds, p.mut = 0.4)
}

