% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.mutator.insertion.R
\name{mutInsertion}
\alias{mutInsertion}
\title{Insertion mutator.}
\usage{
mutInsertion(ind)
}
\arguments{
\item{ind}{[\code{integer}]\cr
Permutation of integers, i.e., vector of integer values.}
}
\value{
[\code{integer}]
}
\description{
The Insertion mutation operator selects a position random and inserts it at
a random position.
}
\seealso{
Other mutators: \code{\link{mutBitflip}},
  \code{\link{mutGauss}}, \code{\link{mutPolynomial}},
  \code{\link{mutScramble}}, \code{\link{mutSwap}},
  \code{\link{mutUniform}}
}
