% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.mutator.inversion.R
\name{mutInversion}
\alias{mutInversion}
\title{Inversion mutator.}
\usage{
mutInversion(ind)
}
\arguments{
\item{ind}{[\code{integer}]\cr
Permutation of integers, i.e., vector of integer values.}
}
\value{
[\code{integer}]
}
\description{
The Inversion mutation operator selects two positions within the chromosome at
random and inverts the elements inbetween.
}
