% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.mutator.jump.R
\name{mutJump}
\alias{mutJump}
\title{Jump mutator.}
\usage{
mutJump(ind)
}
\arguments{
\item{ind}{[\code{integer}]\cr
Permutation of integers, i.e., vector of integer values.}
}
\value{
[\code{integer}]
}
\description{
The jump mutation operator selects two positions within the chromosome at
random, say \eqn{a} and \eqn{b} with \eqn{a < b}. Next, all elements at
positions \eqn{b-1, b-2, ..., a} are shifted to the right by one position
and finally the element at position \eqn{b} is assigned at position \eqn{a}.
}
\seealso{
Other mutators: 
\code{\link{mutBitflip}()},
\code{\link{mutGauss}()},
\code{\link{mutInsertion}()},
\code{\link{mutInversion}()},
\code{\link{mutPolynomial}()},
\code{\link{mutScramble}()},
\code{\link{mutSwap}()},
\code{\link{mutUniform}()}
}
\concept{mutators}
