% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.mutator.scramble.R
\name{mutScramble}
\alias{mutScramble}
\title{Scramble mutator.}
\usage{
mutScramble(ind)
}
\arguments{
\item{ind}{[\code{integer}]\cr
Permutation of integers, i.e., vector of integer values.}
}
\value{
[\code{integer}]
}
\description{
The Scramble mutation operator selects two positions within the chromosome at
random and randomly intermixes the subsequence between these positions.
}
\seealso{
Other mutators: 
\code{\link{mutBitflip}()},
\code{\link{mutGauss}()},
\code{\link{mutInsertion}()},
\code{\link{mutInversion}()},
\code{\link{mutJump}()},
\code{\link{mutPolynomial}()},
\code{\link{mutSwap}()},
\code{\link{mutUniform}()}
}
\concept{mutators}
