% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.mutator.swap.R
\name{mutSwap}
\alias{mutSwap}
\title{Swap mutator.}
\usage{
mutSwap(ind)
}
\arguments{
\item{ind}{[\code{integer}]\cr
Permutation of integers, i.e., vector of integer values.}
}
\value{
[\code{integer}]
}
\description{
Chooses two positions at random and swaps the genes.
}
\seealso{
Other mutators: \code{\link{mutBitflip}},
  \code{\link{mutGauss}}, \code{\link{mutInsertion}},
  \code{\link{mutPolynomial}}, \code{\link{mutScramble}},
  \code{\link{mutUniform}}
}
