% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominates.R
\name{plotFront}
\alias{plotFront}
\title{Plot Pareto-front.}
\usage{
plotFront(x, obj.names = NULL)
}
\arguments{
\item{x}{[\code{matrix} | \code{data.frame}]\cr
Object which contains the non-dominated points.}

\item{obj.names}{[\code{character}]\cr
Optional objectives names.}
}
\value{
[\code{ggplot}] \pkg{ggplot} object.
}
\description{
Plots a scatterplot of non-dominated points in the objective space
utilizing the \pkg{ggplot2} package. The function returns a \code{ggplot} object
which can be furhter modified via additional ggplot layers.
If the passed object is a \code{data.frame}, each line is considered to contain
the fitness values of one individual. Contrary, if a matrix is passed, it is
considered to be passed in ecr2 format, i.e., each column corresponds to one
individual. The matrix is then transposed and converted to a \code{data.frame}.
}
\note{
At the moment only two-dimensional objective spaces are supported.
}
\examples{
matrix
}

