% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFronts.R
\name{plotFronts3d}
\alias{plotFronts3d}
\title{Visualize three-objective Pareto-front approximations.}
\usage{
plotFronts3d(df, obj.cols = c("f1", "f2", "f3"), max.in.col = 4L,
  driver = "scatterplot3d", ...)
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data.frame with columns at least \code{obj.cols}, \dQuote{prob} and \dQuote{algorithm}.}

\item{obj.cols}{[\code{character(>= 3)}]\cr
Column names of the objective functions.
Default is \code{c("f1", "f2", "f3")}.}

\item{max.in.col}{[\code{integer(1)}]\cr
Maximum number of plots to be displayed side by side.
Default is 4.}

\item{...}{[any]\cr
Furhter parameters passed down to plot function.}
}
\value{
Nothing
}
\description{
Given a data frame with the results of (multiple) runs of (multiple)
different three-objective optimization algorithms on (multiple) problem instances
the function generates 3D scatterplots of the obtained Pareto-front approximations.
}
