% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndicatorDistributions.R
\name{plotIndicatorDistribution}
\alias{plotIndicatorDistribution}
\title{Plot distributions of EMOA indicators.}
\usage{
plotIndicatorDistribution(inds, plot.type = "boxplot", facet.type = "grid",
  facet.args = list(), logscale = character())
}
\arguments{
\item{inds}{[\code{data.frame}]\cr
Data frame with columns \dQuote{algorithm}, \dQuote{prob}, \dQuote{repl} and
one additional column per EMOA indicator.}

\item{plot.type}{[\code{character(1)}]\cr
Either \dQuote{boxplot} (the default) for boxplots or \dQuote{violin} for
violin plots.}
}
\value{
[\code{\link[ggplot2]{ggplot}}]
}
\description{
Visualizes of empirical distributions of unary EMOA indicator
based on the results of \code{\link{computeIndicators}}.
}
