% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.default.R
\name{plotStatistics}
\alias{plotStatistics}
\title{Generate line plot of logged statistics.}
\usage{
plotStatistics(x, drop.stats = character(0L))
}
\arguments{
\item{x}{[\code{ecr_statistics} | \code{ecr_logger}]\cr
Logger object or statistics data frame from logger object.}

\item{drop.stats}{[\code{character}]\cr
Names of logged statistics to be dropped.
Default is the empty character, i.e., not to drop any stats.}
}
\description{
Expects a data.frame of logged statistics, e.g., extracted from
a logger object by calling \code{\link{getStatistics}}, and generates a basic
line plot. The plot is generated with the \pkg{ggplot2} package and the ggplot
object is returned.
}
