% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.recombinator.intermediate.R
\name{recIntermediate}
\alias{recIntermediate}
\title{Indermediate recombinator.}
\usage{
recIntermediate(inds)
}
\arguments{
\item{inds}{[\code{list}]\cr
Parents, i.e., list of exactly two numeric vectors of equal length.}
}
\value{
[\code{numeric}] Single offspring.
}
\description{
Intermediate recombination computes the component-wise mean value of the
\code{k} given parents. It is applicable only for float representation.
}
\seealso{
Other recombinators: 
\code{\link{recCrossover}()},
\code{\link{recOX}()},
\code{\link{recPMX}()},
\code{\link{recSBX}()},
\code{\link{recUnifCrossover}()}
}
\concept{recombinators}
