% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.recombinator.pmx.R
\name{recPMX}
\alias{recPMX}
\title{Partially-Mapped-Crossover (PMX) recombinator.}
\usage{
recPMX(inds)
}
\arguments{
\item{inds}{[\code{numeric}]\cr
Parents, i.e., list of exactly two permutations of equal length.}
}
\value{
[\code{ecr_recombinator}]
}
\description{
This recombination operator is specifically designed for permutations.
The operators chooses two cut-points at random and generates two offspring
as follows: a) copy the subsequence of one parent and b) fill the remaining
positions while preserving the order and position of as many genes as possible.
}
\seealso{
Other recombinators: 
\code{\link{recCrossover}()},
\code{\link{recIntermediate}()},
\code{\link{recOX}()},
\code{\link{recSBX}()},
\code{\link{recUnifCrossover}()}
}
\concept{recombinators}
