% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.recombinator.uniform.R
\name{recUnifCrossover}
\alias{recUnifCrossover}
\title{Uniform crossover recombinator.}
\usage{
recUnifCrossover(inds, p = 0.5)
}
\arguments{
\item{inds}{[\code{list}]\cr
Parents, i.e., list of exactly two numeric or binary vectors of equal length.}

\item{p}{[\code{numeric(1)}]\cr
Probability to select gene from parent1.}
}
\value{
[\code{list}]
}
\description{
Produces two child individuals. The i-th gene is from parent1 with probability
\code{p} and from parent2 with probability \code{1-p}.
}
\seealso{
Other recombinators: 
\code{\link{recCrossover}()},
\code{\link{recIntermediate}()},
\code{\link{recOX}()},
\code{\link{recPMX}()},
\code{\link{recSBX}()}
}
\concept{recombinators}
