% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.utils.R
\name{reduceToSingleDataFrame}
\alias{reduceToSingleDataFrame}
\title{Combine multiple Pareto-front approximations into a single data.frame.}
\usage{
reduceToSingleDataFrame(res = list(), group.col.name)
}
\description{
Combine multiple Pareto-front approximations into a single data.frame.
}
