% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerECRObject.R
\name{registerECRObject}
\alias{registerECRObject}
\title{Register object to control object.}
\usage{
registerECRObject(control, name, object)
}
\arguments{
\item{control}{[\code{ecr2_control}]\cr
Control object.}

\item{name}{[\code{character(1)}]\cr
Name to use when object is stored to control object.}

\item{object}{[any]\cr
Object to store.}
}
\value{
[\code{ecr2_control}]
  Modified control object.
}
\description{
Register object to control object.
}

