% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerECRObject.R
\name{registerECROperator}
\alias{registerECROperator}
\title{Helper function to set operator internally.}
\usage{
registerECROperator(control, slot, fun, ...)
}
\arguments{
\item{control}{[\code{ecr_control}]\cr
ECR control object.}

\item{slot}{[\code{character(1)}]\cr
Name of the field in the control object where to store the operator.}

\item{fun}{[\code{function}]\cr
Actual operator. In order to use the various helper functions of ecr one needs
to stick to a simple convention: The first argument of \code{function} should
be the individual to mutate, a list of individuals for recombination or a matrix
of fitness values for recombination. If one does not want to use the corresponding
helpers, e.g., \code{mutate}, the signature of the function does not matter. However,
in this case you are responsable to pass arguments correctly.}

\item{...}{[any]\cr
Further arguments for \code{fun}. These arguments are stored in the control object
and passed on to \code{fun}.}
}
\value{
[\code{ecr_control}]
}
\description{
Helper function to set operator internally.
}

