% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerECRObject.R
\name{registerECRParams}
\alias{registerECRParams}
\title{Register control parameters}
\usage{
registerECRParams(control, ...)
}
\arguments{
\item{control}{[\code{ecr2_control}]\cr
Control object.}

\item{...}{[any]\cr
Further control parameters appended to the control object.}
}
\value{
[\code{ecr2_control}]
}
\description{
Add further parameters which are neccessary for an operator or
several operators. These are passed down automatically to all operator calls
when using, e.g., \code{generateOffspring}.
}

