% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.selector.domNumberPlusHV.R
\name{selDomNumberPlusHV}
\alias{selDomNumberPlusHV}
\title{Modified dominated Hypervolume selector.}
\usage{
selDomNumberPlusHV(fitness, n.select, ref.point)
}
\arguments{
\item{fitness}{[\code{matrix}]\cr
Matrix of fitness values (each column contains the fitness value(s) of one
individual).}

\item{n.select}{[\code{integer(1)}]\cr
Number of elements to select.}

\item{ref.point}{[\code{numeric}]\cr
Reference point for hypervolume computation.}
}
\value{
[\code{integer}] Vector of survivor indizes.
}
\description{
Alternative SMS-EMOA survival selection as proposed in Algorithm 3 of [1]. Performs non-dominated
sorting first. If the number of non-domination levels is at least two the algorithm
drops the individual with the highest number of dominating points (ties are
broken at random) from the last layer. If there is just one non-domination layer,
i.e., all points are non-domianted the method drops the individual with minimal
hypervolume contribution. This selector is the basis of the
S-Metric Selection Evolutionary Multi-Objective Algorithm, termed SMS-EMOA
(see \code{\link{smsemoa}}).
}
\note{
Note that the current implementation expects \code{n.select = ncol(fitness) - 1}
and the selection process quits with an error message if \code{n.select} is greater
than 1.
}
\references{
[1] Beume, Nicola, Boris Naujoks and M. Emmerich. SMS-EMOA: Multiobjective selection
based on dominated hypervolume.” European Journal of Operational Research. 181 (2007): 1653-1669.
}
\seealso{
Other selectors: 
\code{\link{selDomHV}()},
\code{\link{selGreedy}()},
\code{\link{selNondom}()},
\code{\link{selRanking}()},
\code{\link{selRoulette}()},
\code{\link{selSimple}()},
\code{\link{selTournament}()}
}
\concept{selectors}
