% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.selector.greedy.R
\name{selGreedy}
\alias{selGreedy}
\title{Simple selector.}
\usage{
selGreedy(fitness, n.select)
}
\arguments{
\item{fitness}{[\code{matrix}]\cr
Matrix of fitness values (each column contains the fitness value(s) of one
individual).}

\item{n.select}{[\code{integer(1)}]\cr
Number of elements to select.}
}
\value{
[\code{integer}] Vector of survivor indizes.
}
\description{
Sorts the individuals according to their fitness value in increasing order
and selects the best ones.
}
\seealso{
Other selectors: 
\code{\link{selDomHV}()},
\code{\link{selNondom}()},
\code{\link{selRanking}()},
\code{\link{selRoulette}()},
\code{\link{selSimple}()},
\code{\link{selTournament}()}
}
\concept{selectors}
