% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.selector.simple.R
\name{selSimple}
\alias{selSimple}
\title{Simple (naive) selector.}
\usage{
selSimple(fitness, n.select)
}
\arguments{
\item{fitness}{[\code{matrix}]\cr
Matrix of fitness values (each column contains the fitness value(s) of one
individual).}

\item{n.select}{[\code{integer(1)}]\cr
Number of elements to select.}
}
\value{
[\code{setOfIndividuals}]
}
\description{
Just for testing. Actually does not really select, but instead returns a random
sample of \code{ncol(fitness)} indizes.
}
\seealso{
Other selectors: \code{\link{selDomHV}},
  \code{\link{selGreedy}}, \code{\link{selNondom}},
  \code{\link{selRoulette}}, \code{\link{selTournament}}
}
