% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.selector.dominatedHypervolume.R
\name{setupDominatedHypervolumeSelector}
\alias{setupDominatedHypervolumeSelector}
\title{Dominated hypervolume selector.}
\usage{
setupDominatedHypervolumeSelector(ref.point)
}
\arguments{
\item{ref.point}{[\code{numeric}]\cr
Reference point for hypervolume computation.}
}
\value{
[\code{integer}] Vector of survivor indizes.
}
\description{
Performs nondominated sorting and drops the individual from the last front
with minimal hypervolume contribution.
}
\seealso{
Other selectors: \code{\link{setupGreedySelector}},
  \code{\link{setupNondomSelector}},
  \code{\link{setupRouletteWheelSelector}},
  \code{\link{setupSimpleSelector}},
  \code{\link{setupTournamentSelector}}
}
