% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.selector.greedy.R
\name{setupGreedySelector}
\alias{setupGreedySelector}
\title{Simple \dQuote{greedy} selector.}
\usage{
setupGreedySelector()
}
\value{
[\code{setOfIndividuals}]
}
\description{
Sorts the individuals according to their fitness value in increasing order
and selects the best ones.
}
\seealso{
Other selectors: \code{\link{setupDominatedHypervolumeSelector}},
  \code{\link{setupNondomSelector}},
  \code{\link{setupRouletteWheelSelector}},
  \code{\link{setupSimpleSelector}},
  \code{\link{setupTournamentSelector}}
}

