% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.mutator.insertion.R
\name{setupInsertionMutator}
\alias{setupInsertionMutator}
\title{Generator for the Insertion mutation operator.}
\usage{
setupInsertionMutator()
}
\value{
[\code{ecr2_mutator}]
}
\description{
The Insertion mutation operator selects a position random and inserts it at
a random position.
}
\seealso{
Other mutators: \code{\link{setupBitflipMutator}},
  \code{\link{setupGaussMutator}},
  \code{\link{setupPolynomialMutator}},
  \code{\link{setupScrambleMutator}},
  \code{\link{setupSwapMutator}},
  \code{\link{setupUniformMutator}}
}

