% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.recombinator.intermediate.R
\name{setupIntermediateRecombinator}
\alias{setupIntermediateRecombinator}
\title{Generator of the indermediate recombination operator.}
\usage{
setupIntermediateRecombinator(k = 2L)
}
\arguments{
\item{k}{[integer(1)]\cr
Number of parents required for mating. Default is \code{2}.}
}
\value{
[\code{ecr_recombinator}]
}
\description{
Intermediate recombination computes the component-wise mean value of the
\code{k} given parents. It is applicable only for float representation.
}
\seealso{
Other recombinators: \code{\link{setupCrossoverRecombinator}},
  \code{\link{setupNullRecombinator}},
  \code{\link{setupOXRecombinator}},
  \code{\link{setupPMXRecombinator}},
  \code{\link{setupSBXRecombinator}}
}

