% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.selector.nondom.R
\name{setupNondomSelector}
\alias{setupNondomSelector}
\title{Non-dominated sorting selector.}
\usage{
setupNondomSelector()
}
\value{
[\code{setOfIndividuals}]
}
\description{
Applies nondominated sorting of the objective and subsequent crowding distance
criterion to select a subset of individuals. This is the selector used by the
NSGA-II EMOA (see \code{\link{nsga2}}).
}
\seealso{
Other selectors: \code{\link{setupDominatedHypervolumeSelector}},
  \code{\link{setupGreedySelector}},
  \code{\link{setupRouletteWheelSelector}},
  \code{\link{setupSimpleSelector}},
  \code{\link{setupTournamentSelector}}
}

