% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.recombinator.null.R
\name{setupNullRecombinator}
\alias{setupNullRecombinator}
\title{Generator of the \dQuote{null} recombination operator.}
\usage{
setupNullRecombinator()
}
\value{
[\code{ecr2_recombinator}]
}
\description{
This recombinator does not perform any recombination. It simply returns the
first of the passed parents.
}
\seealso{
Other recombinators: \code{\link{setupCrossoverRecombinator}},
  \code{\link{setupIntermediateRecombinator}},
  \code{\link{setupOXRecombinator}},
  \code{\link{setupPMXRecombinator}},
  \code{\link{setupSBXRecombinator}}
}

