% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.generator.random.permutation.R
\name{setupPermutationGenerator}
\alias{setupPermutationGenerator}
\title{Generates a generator object for the initial population.}
\usage{
setupPermutationGenerator(len, set = seq_len(len))
}
\arguments{
\item{len}{[\code{integer(1)}]\cr
Genotype length.}

\item{set}{[\code{vector}]\cr
Permutations will be generated from this set.
Default is \code{1:len}.}
}
\value{
[\code{ecr2_operator}]
}
\description{
This generator generates a population of permutations.
}

