% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.mutator.scramble.R
\name{setupScrambleMutator}
\alias{setupScrambleMutator}
\title{Generator for the Scramble mutation operator.}
\usage{
setupScrambleMutator()
}
\value{
[\code{ecr_mutator}]
}
\description{
The Scramble mutation operator selects two positions within the chromosome at
random and randomly intermixes the subsequence between these positions.
}
\seealso{
Other mutators: \code{\link{setupBitflipMutator}},
  \code{\link{setupGaussMutator}},
  \code{\link{setupInsertionMutator}},
  \code{\link{setupPolynomialMutator}},
  \code{\link{setupSwapMutator}},
  \code{\link{setupUniformMutator}}
}
