% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.selector.simple.R
\name{setupSimpleSelector}
\alias{setupSimpleSelector}
\title{Simple (naive) mating pool generator.}
\usage{
setupSimpleSelector()
}
\value{
[\code{setOfIndividuals}]
}
\description{
Just for testing. Actually does not really select, but instead returns the
entire population to form the mating pool.
}
\seealso{
Other selectors: \code{\link{setupDominatedHypervolumeSelector}},
  \code{\link{setupGreedySelector}},
  \code{\link{setupNondomSelector}},
  \code{\link{setupRouletteWheelSelector}},
  \code{\link{setupTournamentSelector}}
}

