% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.mutator.swap.R
\name{setupSwapMutator}
\alias{setupSwapMutator}
\title{Generator for the Swap mutation operator.}
\usage{
setupSwapMutator()
}
\value{
[\code{ecr_mutator}]
}
\description{
Chooses two positions at random and swaps the genes.
}
\seealso{
Other mutators: \code{\link{setupBitflipMutator}},
  \code{\link{setupGaussMutator}},
  \code{\link{setupInsertionMutator}},
  \code{\link{setupPolynomialMutator}},
  \code{\link{setupScrambleMutator}},
  \code{\link{setupUniformMutator}}
}
