% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.default.R
\name{toGG}
\alias{toGG}
\title{Transform to long format.}
\usage{
toGG(x, drop.stats = character(0L))
}
\arguments{
\item{x}{[\code{ecr_statistics} | \code{ecr_logger}]\cr
Logger object or statistics data frame from logger object.}

\item{drop.stats}{[\code{character}]\cr
Names of logged statistics to be dropped.
Default is the empty character, i.e., not to drop any stats.}
}
\value{
[\code{data.frame}]
}
\description{
Transform the data.frame of logged statistics from wide to
\pkg{ggplot2}-friendly long format.
}
