% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PA.EMOA.toLatex.R
\name{toLatex}
\alias{toLatex}
\title{Export results of statistical tests to LaTeX table(s).}
\usage{
toLatex(stats, probs = NULL, inds = NULL, type = "by.instance",
  cell.formatter = NULL)
}
\arguments{
\item{stats}{[\code{list}]\cr
Named list of list as returned by \code{\link{test}}.}

\item{probs}{[\code{character}]\cr
Filtering: vector of problem instances. This way one can restrict the
size of the table(s).
Defaults to all problems available in \code{stats}.}

\item{inds}{[\code{character}]\cr
Names of the indicators to consider.
Defaults to all indicators available in \code{stats}.}

\item{type}{[\code{character(1)}]\cr
Type of tables. At the moment only option \dQuote{by.instance} is available.
I.e., a separate LaTeX-table is generated for each instance specified via \code{probs}.}

\item{cell.formatter}{[\code{function(cell, ...)}]\cr
Function which is used to format table cells. This function is applied to each
table cell and may be used to customize the output. Default is \code{niceCellFormater}.}
}
\value{
[\code{list}] Named list of strings (LaTeX tables). Names correspond to the
  selected problem instances in \code{probs}.
}
\description{
Returns high-quality LaTeX-tables of the test results of
statistical tests performed with function \code{\link{test}}
on per-instance basis. I.e., a table is returned for each instances combining
the results of different indicators.
}
\seealso{
Other EMOA performance assessment tools: \code{\link{approximateNadirPoint}},
  \code{\link{computeDominanceRanking}},
  \code{\link{emoaIndEps}},
  \code{\link{makeEMOAIndicator}},
  \code{\link{niceCellFormater}}, \code{\link{normalize}},
  \code{\link{plotDistribution}}, \code{\link{plotFront}},
  \code{\link{plotScatter2d}}, \code{\link{plotScatter3d}},
  \code{\link{test}}
}
