% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.utils.R
\name{toParetoDf}
\alias{toParetoDf}
\title{Convert matrix to Pareto front data frame.}
\usage{
toParetoDf(x, filter.dups = FALSE)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Matrix.}

\item{filter.dups}{[\code{logical(1)}]\cr
Shall duplicates be removed?
Default is \code{FALSE}.}
}
\value{
[\code{data.frame}]
}
\description{
Inside ecr EMOA algorithms the fitness is maintained in an \eqn{(o, n)} matrix
where \eqn{o} is the number of objectives and \eqn{n} is the number of individuals.
This function basically transposes such a matrix and converts it into a data frame.
}
