% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformFitness.R
\name{transformFitness}
\alias{transformFitness}
\title{Fitness transformation / scaling.}
\usage{
transformFitness(fitness, task, selector)
}
\arguments{
\item{fitness}{[matrix]
Matrix of fitness values with the fitness vector of individual i in the i-th
column.}

\item{task}{[ecr_optimization_task]
Optimization task.}

\item{selector}{[ecr_selector] 
Selector object.}
}
\value{
[matrix] Transformed / scaled fitness matrix.
}
\description{
Some selectors support maximization only, e.g., roulette wheel selector, or
minimization (most others). This function computes a factor from {-1, 1} for
each objective to match supported selector optimization directions and
the actual objectives of the task.
}
