% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wrapChildren}
\alias{wrapChildren}
\title{Wrap the individuals constructed by a recombination operator.}
\usage{
wrapChildren(...)
}
\arguments{
\item{...}{[any]\cr
Individuals.}
}
\value{
[\code{list}] List of individuals.
}
\description{
Should be used if the recombinator returns multiple children.
}

