% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initECRControl.R
\name{initECRControl}
\alias{initECRControl}
\title{Control object generator.}
\usage{
initECRControl(fitness.fun, n.objectives = NULL, minimize = NULL)
}
\arguments{
\item{fitness.fun}{[\code{function}]\cr
The fitness function.}

\item{n.objectives}{[\code{integer(1)}]\cr
Number of objectives of \code{obj.fun}.
Optional if \code{obj.fun} is a benchmark function from package \pkg{smoof}.}

\item{minimize}{[\code{logical(n.objectives)}]\cr
Logical vector with ith entry \code{TRUE} if the ith objective of \code{fitness.fun}
shall be minimized. If a single logical is passed, it is assumed to be valid
for each objective.}
}
\value{
[\code{ecr_control}]
}
\description{
The control object keeps information on the objective function and a set of
evolutionary components, i.e., operators.
}
