% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.default.R
\name{setupECRDefaultMonitor}
\alias{setupECRDefaultMonitor}
\title{Default monitor.}
\usage{
setupECRDefaultMonitor(step = 10L)
}
\arguments{
\item{step}{[\code{integer(1)}]\cr
Number of steps of the EA between monitoring.
Default is 10.}
}
\value{
[\code{ecr_monitor}]
}
\description{
Default monitor object that outputs messages to the console
based on a default logger (see \code{initLogger}).
}
