% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.mutator.polynomial.R
\name{setupPolynomialMutator}
\alias{setupPolynomialMutator}
\title{Generator of the polynomial mutation operator.}
\usage{
setupPolynomialMutator(p = 0.2, eta = 10)
}
\arguments{
\item{p}{[\code{numeric(1)}]\cr
Probability of mutation of each gene.}

\item{eta}{[\code{numeric(1)}\cr
Distance parameter of the mutation distribution.}
}
\value{
[\code{ecr_mutator}]
}
\description{
Performs an polynomial mutation as used in the SMS-EMOA algorithm.
}
\seealso{
Other mutators: \code{\link{setupBitflipMutator}},
  \code{\link{setupGaussMutator}},
  \code{\link{setupInsertionMutator}},
  \code{\link{setupScrambleMutator}},
  \code{\link{setupSwapMutator}},
  \code{\link{setupUniformMutator}}
}

