% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminators.R
\name{stoppingConditions}
\alias{stoppingConditions}
\alias{stopOnEvals}
\alias{stopOnIters}
\title{Stopping condition: maximum number of function evaluations.}
\usage{
stopOnEvals(max.evals = NULL)

stopOnIters(max.iter = NULL)
}
\arguments{
\item{max.evals}{[\code{integer(1)}]\cr
Maximal number of function evaluations. Default ist \code{Inf}.}

\item{max.iter}{[\code{integer(1)}]\cr
Maximal number of iterations. Default ist \code{Inf}.}
}
\value{
[\code{ecr_terminator}]
}
\description{
Stop the EA after at most \code{max.evals} evaluations of the fitness function.
}
