% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.default.R
\name{updateLogger}
\alias{updateLogger}
\title{Update the log.}
\usage{
updateLogger(log, population, fitness = NULL, n.evals, ...)
}
\arguments{
\item{log}{[\code{ecr_logger}]\cr
The log generated by \code{initLogger}.}

\item{population}{[\code{list}]\cr
List of individuals.}

\item{fitness}{[\code{matrix}]\cr
Optional matrix of fitness values (each column contains the fitness value(s) for
one individual) of \code{population}. If no matrix is passed and the log shall
store information of the fitness, each individual needs to have an attribute fitness.}

\item{n.evals}{[\code{integer(1)}]\cr
Number of fitness function evaluations performed in the last generation.}

\item{...}{[any]\cr
Furhter arguments. Not used at the moment.}
}
\description{
This function modifies the log in-place, i.e., without making
copies.
}
\seealso{
Other logging: \code{\link{getPopulations}},
  \code{\link{getStatistics}}, \code{\link{initLogger}}
}
